Linear system and Pendulum Rendering (#39)
Added rendering functionality for the Linear system and Pendulum on a
cart environments.
In particular note:
- The new dependencies that must be installed if rendering should be
done.
- Integration in the Linear System SAC script is done, but we should
still talk about whether and if yes, how we want rendering built-in into
trainer_base.
- Rendering queries the parameters from the parameters_dict given to the
mpc in the constructor, but queries the mpc when calculating, e.g., one
step of the dynamics. I have opened a ticket for this.
- Live rendering ("human" mode) for pendulum on cart is not tested, but
it is very likely it will work.
- I upped the noise in the pendulum environment, because it was very
weak.
- Testing is basically done by running rollouts in the environments and
manually looking at the videos. If the tests run, i.e., nothing bad
happens while creating the frames/videos, the tests pass (even when the
videos are bad). But I cant think of a better way to test rendering (at
best, we could write tests for the geometric calculations of the
rendering_utils, but not sure whether that is worth the hassle, when you
can just look at it? :D )