Adding a drone example (#35)
The drone environment involves stabilizing a quadrotor from a random
position on a sphere at the origin. The MPC horizon is kept very short.
Two variants of the task are contained: 1. varying the mass of the drone
in a closed-loop and 2. varying the terminal reference in the
closed-loop.
An evaluation script is added to test the MPC implementation.

---------

Co-authored-by: Jasper Hoffmann <hoffmaja@informatik.uni-freiburg.de>
Co-authored-by: dirk <dirk.p.reinhardt@ntnu.no>
Co-authored-by: JasperHoffmann <41264755+JasperHoffmann@users.noreply.github.com>