Logging: minor fixes (#100)
This PR proposes the following changes:
1. fix accessing `wandb_init_kwargs`, e.g., when setting custom run
names
2. add an initial policy validation (perhaps the initial validation
score does not start at zero)
3. move closing of the Logger before the `return` command, otherwise it
does not close after the run in some cases

Also, I noticed that when `val_report_score` is `"final"`, the best
score is returned instead of the last one.
Should a `"best"` option be added and the rest modified accordingly?

---------

Co-authored-by: JasperHoffmann <41264755+JasperHoffmann@users.noreply.github.com>